pyproject.toml
[tool.poetry]
name = "coins_wallet_service"
version = "0.1.0"
description = ""
authors = ["Bulat Khasanov <afti@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
djangorestframework = "^3.9"
django-extensions = "^2.1"
psycopg2-binary = "^2.8"

[tool.poetry.dev-dependencies]
pytest = "^5.0"
pytest-django = "^3.5"
isort = "^4.3"
black = {version = "^18.3-alpha.0",allows-prereleases = true}
sphinx = "^2.1"
ipython = "^7.6"
django_debug_toolbar = "^2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
